Restructure packaging for Voucher
COAND-497
