+++ voucher/src/test/java/com/adyen/checkout/voucher/VoucherComponentTest.kt
@@ -19,6 +19,8 @@ import com.adyen.checkout.components.test.TestComponentViewType
 import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.test.TestDispatcherExtension
 import com.adyen.checkout.test.extensions.invokeOnCleared
+import com.adyen.checkout.voucher.internal.ui.VoucherComponentViewType
+import com.adyen.checkout.voucher.internal.ui.VoucherDelegate
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.runTest
